app/requirements.txt
asgiref==3.8.1
attrs==25.1.0
black==24.8.0
click==8.1.8
Django==3.2.16
django-ckeditor==6.7.2
django-cors-headers==4.4.0
django-js-asset==2.2.0
django-json-widget==2.0.1
django-nested-admin==4.1.1
djangorestframework==3.14.0
importlib_resources==6.4.5
isort==5.13.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
psycopg2==2.9.5
python-monkey-business==1.1.0
pytz==2022.6
referencing==0.35.1
rpds-py==0.20.1
slugify==0.0.1
sqlparse==0.4.3
tomli==2.3.0
typing_extensions==4.12.2
zipp==3.20.2
